+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.2.0
+
+* Breaking change: AUDIO_SESSION_MICROPHONE=0 by default on iOS.
+* Migrate to Kotlin.
+* Bump min flutter version to 3.27.0, AGP to 8.5.2.
+
 ## 0.1.25
 
 * Fix SwiftPM support on macOS.